requirements.txt
# python
langchain
tweepy
python-dotenv
langchain_groq
